+++ src/api/FinancialHub.Core.Services.NUnitTests/Validators/BalanceValidatorTests.cs
@@ -1,6 +1,6 @@
-﻿using FinancialHub.Core.WebApi.Validators;
+﻿using FinancialHub.Core.Services.Validators;
 
-namespace FinancialHub.Core.WebApi.NUnitTests.Validators
+namespace FinancialHub.Core.Services.NUnitTests.Validators
 {
     public class BalanceValidatorTests
     {
